QnA Assist/QnA Assist-Snowflake DB/requirements.txt
snowflake-snowpark-python
snowflake-connector-python
snowflake-sqlalchemy
streamlit
SQLAlchemy
python-dotenv
jsonschema
jsonschema-specifications
cryptography
fastapi
openai
numpy
pandas
plotly
pyarrow
scikit-learn
pmdarima
matplotlib
requests
azure.storage.blob
azure-search-documents
azure-core
azure
uuid